requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
cryptography==3.4.7
Django==3.2.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
pycparser==2.20
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
